+++ resources/views/backend/dashboard.blade.php
@@ -14,7 +14,7 @@
                 <div>
                     <p class="text-slate-600 dark:text-dark-400 text-sm font-medium leading-4">Total Links</p>
                     <p class="text-2xl font-bold text-slate-700 dark:text-dark-100 inline-flex items-center space-x-2">
-                        {{ n_abb($url->authUserLinks(auth()->id())) }}
+                        {{ n_abb($linkService->authUserLinks(auth()->id())) }}
                     </p>
                 </div>
             </div>
